requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Configuration management
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# Google API client
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Google Cloud Speech-to-Text
google-cloud-speech==2.33.0

# Utility
python-multipart==0.0.9

# Audio processing
librosa==0.10.2
soundfile==0.12.1

# WebSocket support
websockets==12.0
